opt-rebalance/requirements.txt
numpy>=1.21.1
scipy>=1.7.1
jax>=0.2.20
jaxlib>=0.1.71
eth-brownie>=1.17.0
